Move Snapshot class from bootstrap to main module
Move CaptureContext as Top level class remaingin in bootstrap
Remove Sink interface from DebuggerContext and use it directly from
DebuggerAgent
Remove ProbeDiagnostic, not used
Move ProbeLocation and EvaluationError as top level class
Rename SnapshotSerializer as ValueSerializer and remove
serializeSnapshot method. Snapshot are serialized directly from
DebuggerAgent's JsonSnapshotSerializer
