+++ dd-java-agent/agent-debugger/debugger-bootstrap/src/main/java/datadog/trace/bootstrap/debugger/EvaluationError.java
@@ -1,0 +1,20 @@
+package datadog.trace.bootstrap.debugger;
+
+/** Store evaluation errors from expressions (probe conditions, log template, metric values, ...) */
+public class EvaluationError {
+  private final String expr;
+  private final String message;
+
+  public EvaluationError(String expr, String message) {
+    this.expr = expr;
+    this.message = message;
+  }
+
+  public String getExpr() {
+    return expr;
+  }
+
+  public String getMessage() {
+    return message;
+  }
+}
