+++ dd-java-agent/agent-debugger/debugger-el/src/test/java/com/datadog/debugger/el/ExpressionTest.java
@@ -8,7 +8,7 @@ import com.datadog.debugger.el.expressions.IsEmptyExpression;
 import com.datadog.debugger.el.values.BooleanValue;
 import com.datadog.debugger.el.values.NumericValue;
 import com.datadog.debugger.el.values.StringValue;
-import datadog.trace.bootstrap.debugger.Snapshot;
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.el.ValueReferenceResolver;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
@@ -21,7 +21,7 @@ class ExpressionTest {
   @MethodSource("literalExpressions")
   void testLiteralExpressions(Literal<?> literal, Object expectedValue) {
     Value<?> value1 = literal.evaluate(null);
-    Value<?> value2 = literal.evaluate(new Snapshot.CapturedContext());
+    Value<?> value2 = literal.evaluate(new CapturedContext());
 
     assertNotNull(value1);
     assertNotNull(value2);
@@ -43,7 +43,7 @@ class ExpressionTest {
     StringValue string = new StringValue("Hello World");
     StringValue emptyString = new StringValue("");
 
-    ValueReferenceResolver resolver = new Snapshot.CapturedContext();
+    ValueReferenceResolver resolver = new CapturedContext();
     IsEmptyExpression isEmpty1 = new IsEmptyExpression(string);
     IsEmptyExpression isEmpty2 = new IsEmptyExpression(emptyString);
