+++ dd-java-agent/agent-debugger/debugger-el/src/test/java/com/datadog/debugger/el/RefResolverHelper.java
@@ -1,6 +1,6 @@
 package com.datadog.debugger.el;
 
-import datadog.trace.bootstrap.debugger.Snapshot;
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.el.ValueReferenceResolver;
 import java.lang.reflect.Field;
 import java.util.*;
@@ -14,47 +14,48 @@ public class RefResolverHelper {
       fields.addAll(Arrays.asList(clazz.getDeclaredFields()));
       clazz = clazz.getSuperclass();
     }
-    Snapshot.CapturedValue[] fieldValues = new Snapshot.CapturedValue[fields.size()];
+    CapturedContext.CapturedValue[] fieldValues = new CapturedContext.CapturedValue[fields.size()];
     int index = 0;
     for (Field field : fields) {
       try {
         field.setAccessible(true);
         fieldValues[index++] =
-            Snapshot.CapturedValue.of(
+            CapturedContext.CapturedValue.of(
                 field.getName(), field.getType().getTypeName(), field.get(instance));
       } catch (Exception ex) {
         ex.printStackTrace();
       }
     }
-    return new Snapshot.CapturedContext(null, null, null, null, fieldValues);
+    return new CapturedContext(null, null, null, null, fieldValues);
   }
 
   public static ValueReferenceResolver createResolver(
       Map<String, Object> args, Map<String, Object> locals, Map<String, Object> fields) {
-    Snapshot.CapturedValue[] argValues = null;
+    CapturedContext.CapturedValue[] argValues = null;
     if (args != null) {
-      argValues = new Snapshot.CapturedValue[args.size()];
+      argValues = new CapturedContext.CapturedValue[args.size()];
       fillValues(args, argValues);
     }
-    Snapshot.CapturedValue[] localValues = null;
+    CapturedContext.CapturedValue[] localValues = null;
     if (locals != null) {
-      localValues = new Snapshot.CapturedValue[locals.size()];
+      localValues = new CapturedContext.CapturedValue[locals.size()];
       fillValues(locals, localValues);
     }
-    Snapshot.CapturedValue[] fieldValues = null;
+    CapturedContext.CapturedValue[] fieldValues = null;
     if (fields != null) {
-      fieldValues = new Snapshot.CapturedValue[fields.size()];
+      fieldValues = new CapturedContext.CapturedValue[fields.size()];
       fillValues(fields, fieldValues);
     }
-    return new Snapshot.CapturedContext(argValues, localValues, null, null, fieldValues);
+    return new CapturedContext(argValues, localValues, null, null, fieldValues);
   }
 
-  private static void fillValues(Map<String, Object> fields, Snapshot.CapturedValue[] fieldValues) {
+  private static void fillValues(
+      Map<String, Object> fields, CapturedContext.CapturedValue[] fieldValues) {
     int index = 0;
     for (Map.Entry<String, Object> entry : fields.entrySet()) {
       Object value = entry.getValue();
       fieldValues[index++] =
-          Snapshot.CapturedValue.of(
+          CapturedContext.CapturedValue.of(
               entry.getKey(),
               value != null ? value.getClass().getTypeName() : Object.class.getTypeName(),
               value);
