+++ dd-java-agent/agent-debugger/src/main/java/com/datadog/debugger/instrumentation/Types.java
@@ -17,11 +17,11 @@ import static org.objectweb.asm.Opcodes.LASTORE;
 import static org.objectweb.asm.Opcodes.SALOAD;
 import static org.objectweb.asm.Opcodes.SASTORE;
 
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.CorrelationAccess;
 import datadog.trace.bootstrap.debugger.DebuggerContext;
 import datadog.trace.bootstrap.debugger.DebuggerSpan;
 import datadog.trace.bootstrap.debugger.MethodLocation;
-import datadog.trace.bootstrap.debugger.Snapshot;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.StringTokenizer;
@@ -40,10 +40,10 @@ public final class Types {
   public static final Type OBJECT_TYPE = Type.getType(Object.class);
   public static final Type MAP_TYPE = Type.getType(Map.class);
   public static final Type HASHMAP_TYPE = Type.getType(HashMap.class);
-  public static final Type SNAPSHOT_TYPE = Type.getType(Snapshot.class);
-  public static final Type CAPTURED_VALUE = Type.getType(Snapshot.CapturedValue.class);
-  public static final Type CAPTURED_CONTEXT_TYPE = Type.getType(Snapshot.CapturedContext.class);
-  public static final Type CAPTURE_THROWABLE_TYPE = Type.getType(Snapshot.CapturedThrowable.class);
+  public static final Type CAPTURED_VALUE = Type.getType(CapturedContext.CapturedValue.class);
+  public static final Type CAPTURED_CONTEXT_TYPE = Type.getType(CapturedContext.class);
+  public static final Type CAPTURE_THROWABLE_TYPE =
+      Type.getType(CapturedContext.CapturedThrowable.class);
   public static final Type THROWABLE_TYPE = Type.getType(Throwable.class);
   public static final Type CORRELATION_ACCESS_TYPE = Type.getType(CorrelationAccess.class);
   public static final Type DEBUGGER_CONTEXT_TYPE = Type.getType(DebuggerContext.class);
