+++ dd-java-agent/agent-debugger/src/main/java/com/datadog/debugger/probe/LogProbe.java
@@ -2,23 +2,27 @@ package com.datadog.debugger.probe;
 
 import static datadog.trace.bootstrap.debugger.util.TimeoutChecker.DEFAULT_TIME_OUT;
 
+import com.datadog.debugger.agent.DebuggerAgent;
 import com.datadog.debugger.agent.Generated;
 import com.datadog.debugger.agent.LogMessageTemplateBuilder;
 import com.datadog.debugger.el.EvaluationException;
 import com.datadog.debugger.el.ProbeCondition;
 import com.datadog.debugger.el.ValueScript;
+import com.datadog.debugger.instrumentation.DiagnosticMessage;
 import com.datadog.debugger.instrumentation.LogInstrumentor;
+import com.datadog.debugger.sink.Sink;
+import com.datadog.debugger.sink.Snapshot;
 import com.squareup.moshi.Json;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.DebuggerContext;
-import datadog.trace.bootstrap.debugger.DiagnosticMessage;
+import datadog.trace.bootstrap.debugger.EvaluationError;
 import datadog.trace.bootstrap.debugger.Limits;
 import datadog.trace.bootstrap.debugger.MethodLocation;
 import datadog.trace.bootstrap.debugger.ProbeId;
 import datadog.trace.bootstrap.debugger.ProbeRateLimiter;
-import datadog.trace.bootstrap.debugger.Snapshot;
 import datadog.trace.bootstrap.debugger.util.TimeoutChecker;
 import java.io.IOException;
 import java.util.Arrays;
@@ -346,12 +350,12 @@ public class LogProbe extends ProbeDefinition {
   }
 
   @Override
-  public void evaluate(Snapshot.CapturedContext context, Snapshot.CapturedContext.Status status) {
+  public void evaluate(CapturedContext context, CapturedContext.Status status) {
     status.setCondition(evaluateCondition(context, status));
-    Snapshot.CapturedThrowable throwable = context.getThrowable();
+    CapturedContext.CapturedThrowable throwable = context.getThrowable();
     if (status.hasConditionErrors() && throwable != null) {
       status.addError(
-          new Snapshot.EvaluationError(
+          new EvaluationError(
               "uncaught exception", throwable.getType() + ": " + throwable.getMessage()));
     }
     if (status.getCondition()) {
@@ -360,8 +364,7 @@ public class LogProbe extends ProbeDefinition {
     }
   }
 
-  private boolean evaluateCondition(
-      Snapshot.CapturedContext capture, Snapshot.CapturedContext.Status status) {
+  private boolean evaluateCondition(CapturedContext capture, CapturedContext.Status status) {
     if (probeCondition == null) {
       return true;
     }
@@ -372,7 +375,7 @@ public class LogProbe extends ProbeDefinition {
       }
     } catch (EvaluationException ex) {
       LOGGER.debug("Evaluation error: ", ex);
-      status.addError(new Snapshot.EvaluationError(ex.getExpr(), ex.getMessage()));
+      status.addError(new EvaluationError(ex.getExpr(), ex.getMessage()));
       status.setConditionErrors(true);
       return false;
     } finally {
@@ -384,11 +387,11 @@ public class LogProbe extends ProbeDefinition {
 
   @Override
   public void commit(
-      Snapshot.CapturedContext entryContext,
-      Snapshot.CapturedContext exitContext,
-      List<Snapshot.CapturedThrowable> caughtExceptions) {
-    Snapshot.CapturedContext.Status entryStatus = entryContext.getStatus(id);
-    Snapshot.CapturedContext.Status exitStatus = exitContext.getStatus(id);
+      CapturedContext entryContext,
+      CapturedContext exitContext,
+      List<CapturedContext.CapturedThrowable> caughtExceptions) {
+    CapturedContext.Status entryStatus = entryContext.getStatus(id);
+    CapturedContext.Status exitStatus = exitContext.getStatus(id);
     String message = null;
     switch (evaluateAt) {
       case ENTRY:
@@ -399,13 +402,14 @@ public class LogProbe extends ProbeDefinition {
         message = exitStatus.getMessage();
         break;
     }
+    Sink sink = DebuggerAgent.getSink();
     boolean shouldCommit = false;
     Snapshot snapshot = new Snapshot(Thread.currentThread(), this);
     if (entryStatus.shouldSend() && exitStatus.shouldSend()) {
       // only rate limit if a condition is defined
       if (probeCondition != null) {
         if (!ProbeRateLimiter.tryProbe(id)) {
-          DebuggerContext.skipSnapshot(id, DebuggerContext.SkipCause.RATE);
+          sink.skipSnapshot(id, DebuggerContext.SkipCause.RATE);
           return;
         }
       }
@@ -435,13 +439,13 @@ public class LogProbe extends ProbeDefinition {
       shouldCommit = true;
     }
     if (shouldCommit) {
-      commitSnapshot(snapshot);
+      commitSnapshot(snapshot, sink);
     } else {
-      DebuggerContext.skipSnapshot(id, DebuggerContext.SkipCause.CONDITION);
+      sink.skipSnapshot(id, DebuggerContext.SkipCause.CONDITION);
     }
   }
 
-  private void commitSnapshot(Snapshot snapshot) {
+  private void commitSnapshot(Snapshot snapshot, Sink sink) {
     /*
      * Record stack trace having the caller of this method as 'top' frame.
      * For this it is necessary to discard:
@@ -452,22 +456,23 @@ public class LogProbe extends ProbeDefinition {
      * - DebuggerContext.commit() or DebuggerContext.evalAndCommit()
      */
     snapshot.recordStackTrace(5);
-    DebuggerContext.addSnapshot(snapshot);
+    sink.addSnapshot(snapshot);
   }
 
   @Override
-  public void commit(Snapshot.CapturedContext lineContext, int line) {
-    Snapshot.CapturedContext.Status status = lineContext.getStatus(id);
+  public void commit(CapturedContext lineContext, int line) {
+    CapturedContext.Status status = lineContext.getStatus(id);
     if (status == null) {
       return;
     }
+    Sink sink = DebuggerAgent.getSink();
     Snapshot snapshot = new Snapshot(Thread.currentThread(), this);
     boolean shouldCommit = false;
     if (status.shouldSend()) {
       // only rate limit if a condition is defined
       if (probeCondition != null) {
         if (!ProbeRateLimiter.tryProbe(id)) {
-          DebuggerContext.skipSnapshot(id, DebuggerContext.SkipCause.RATE);
+          sink.skipSnapshot(id, DebuggerContext.SkipCause.RATE);
           return;
         }
       }
@@ -484,10 +489,10 @@ public class LogProbe extends ProbeDefinition {
     if (shouldCommit) {
       // freeze context just before commit because line probes have only one context
       lineContext.freeze(new TimeoutChecker(DEFAULT_TIME_OUT));
-      commitSnapshot(snapshot);
+      commitSnapshot(snapshot, sink);
       return;
     }
-    DebuggerContext.skipSnapshot(id, DebuggerContext.SkipCause.CONDITION);
+    sink.skipSnapshot(id, DebuggerContext.SkipCause.CONDITION);
   }
 
   @Override
