+++ dd-java-agent/agent-debugger/src/main/java/com/datadog/debugger/util/MoshiSnapshotHelper.java
@@ -1,13 +1,15 @@
 package com.datadog.debugger.util;
 
+import com.datadog.debugger.sink.Snapshot;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.Limits;
 import datadog.trace.bootstrap.debugger.ProbeImplementation;
-import datadog.trace.bootstrap.debugger.Snapshot;
+import datadog.trace.bootstrap.debugger.ProbeLocation;
 import datadog.trace.bootstrap.debugger.util.TimeoutChecker;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -59,10 +61,10 @@ public class MoshiSnapshotHelper {
         return new CapturesAdapter(
             moshi, new CapturedContextAdapter(moshi, new CapturedValueAdapter()));
       }
-      if (Types.equals(type, Snapshot.CapturedValue.class)) {
+      if (Types.equals(type, CapturedContext.CapturedValue.class)) {
         return new CapturedValueAdapter();
       }
-      if (Types.equals(type, Snapshot.CapturedContext.class)) {
+      if (Types.equals(type, CapturedContext.class)) {
         return new CapturedContextAdapter(moshi, new CapturedValueAdapter());
       }
       if (Types.equals(type, ProbeImplementation.class)) {
@@ -73,18 +75,18 @@ public class MoshiSnapshotHelper {
   }
 
   public static class CapturesAdapter extends JsonAdapter<Snapshot.Captures> {
-    protected final JsonAdapter<Snapshot.CapturedContext> capturedContextAdapter;
-    protected final JsonAdapter<Map<Integer, Snapshot.CapturedContext>> linesAdapter;
-    protected final JsonAdapter<List<Snapshot.CapturedThrowable>> caughtExceptionsAdapter;
+    protected final JsonAdapter<CapturedContext> capturedContextAdapter;
+    protected final JsonAdapter<Map<Integer, CapturedContext>> linesAdapter;
+    protected final JsonAdapter<List<CapturedContext.CapturedThrowable>> caughtExceptionsAdapter;
 
-    public CapturesAdapter(
-        Moshi moshi, JsonAdapter<Snapshot.CapturedContext> capturedContextAdapter) {
+    public CapturesAdapter(Moshi moshi, JsonAdapter<CapturedContext> capturedContextAdapter) {
       this.capturedContextAdapter = capturedContextAdapter;
       linesAdapter =
           moshi.adapter(
-              Types.newParameterizedType(Map.class, Integer.class, Snapshot.CapturedContext.class));
+              Types.newParameterizedType(Map.class, Integer.class, CapturedContext.class));
       caughtExceptionsAdapter =
-          moshi.adapter(Types.newParameterizedType(List.class, Snapshot.CapturedThrowable.class));
+          moshi.adapter(
+              Types.newParameterizedType(List.class, CapturedContext.CapturedThrowable.class));
     }
 
     @Override
@@ -114,18 +116,18 @@ public class MoshiSnapshotHelper {
     }
   }
 
-  public static class CapturedContextAdapter extends JsonAdapter<Snapshot.CapturedContext> {
-    protected final JsonAdapter<Snapshot.CapturedThrowable> throwableAdapter;
-    protected final JsonAdapter<Snapshot.CapturedValue> valueAdapter;
+  public static class CapturedContextAdapter extends JsonAdapter<CapturedContext> {
+    protected final JsonAdapter<CapturedContext.CapturedThrowable> throwableAdapter;
+    protected final JsonAdapter<CapturedContext.CapturedValue> valueAdapter;
 
-    public CapturedContextAdapter(Moshi moshi, JsonAdapter<Snapshot.CapturedValue> valueAdapter) {
+    public CapturedContextAdapter(
+        Moshi moshi, JsonAdapter<CapturedContext.CapturedValue> valueAdapter) {
       this.valueAdapter = valueAdapter;
-      this.throwableAdapter = moshi.adapter(Snapshot.CapturedThrowable.class);
+      this.throwableAdapter = moshi.adapter(CapturedContext.CapturedThrowable.class);
     }
 
     @Override
-    public void toJson(JsonWriter jsonWriter, Snapshot.CapturedContext capturedContext)
-        throws IOException {
+    public void toJson(JsonWriter jsonWriter, CapturedContext capturedContext) throws IOException {
       if (capturedContext == null) {
         jsonWriter.nullValue();
         return;
@@ -225,7 +227,7 @@ public class MoshiSnapshotHelper {
     /** @return true if all items where serialized or whether we reach the max field count */
     private SerializationResult toJsonCapturedValues(
         JsonWriter jsonWriter,
-        Map<String, Snapshot.CapturedValue> map,
+        Map<String, CapturedContext.CapturedValue> map,
         Limits limits,
         TimeoutChecker timeoutChecker)
         throws IOException {
@@ -233,7 +235,7 @@ public class MoshiSnapshotHelper {
         return SerializationResult.OK;
       }
       int count = 0;
-      for (Map.Entry<String, Snapshot.CapturedValue> entry : map.entrySet()) {
+      for (Map.Entry<String, CapturedContext.CapturedValue> entry : map.entrySet()) {
         if (count >= limits.maxFieldCount) {
           return SerializationResult.FIELD_COUNT;
         }
@@ -241,7 +243,7 @@ public class MoshiSnapshotHelper {
           return SerializationResult.TIMEOUT;
         }
         jsonWriter.name(entry.getKey());
-        Snapshot.CapturedValue capturedValue = entry.getValue();
+        CapturedContext.CapturedValue capturedValue = entry.getValue();
         jsonWriter.value(
             Okio.buffer(
                 Okio.source(
@@ -253,15 +255,15 @@ public class MoshiSnapshotHelper {
     }
 
     @Override
-    public Snapshot.CapturedContext fromJson(JsonReader reader) throws IOException {
+    public CapturedContext fromJson(JsonReader reader) throws IOException {
       // Only used in test, see MoshiSnapshotTestHelper
       throw new IllegalStateException("Should not reach this code.");
     }
   }
 
-  public static class CapturedValueAdapter extends JsonAdapter<Snapshot.CapturedValue> {
+  public static class CapturedValueAdapter extends JsonAdapter<CapturedContext.CapturedValue> {
     @Override
-    public void toJson(JsonWriter jsonWriter, Snapshot.CapturedValue capturedValue)
+    public void toJson(JsonWriter jsonWriter, CapturedContext.CapturedValue capturedValue)
         throws IOException {
       if (capturedValue == null) {
         jsonWriter.nullValue();
@@ -284,7 +286,7 @@ public class MoshiSnapshotHelper {
     }
 
     @Override
-    public Snapshot.CapturedValue fromJson(JsonReader reader) throws IOException {
+    public CapturedContext.CapturedValue fromJson(JsonReader reader) throws IOException {
       // Only used in test, see MoshiSnapshotTestHelper
       throw new IllegalStateException("Should not reach this code.");
     }
@@ -470,10 +472,10 @@ public class MoshiSnapshotHelper {
   }
 
   public static class ProbeDetailsAdapter extends JsonAdapter<ProbeImplementation> {
-    protected final JsonAdapter<Snapshot.ProbeLocation> probeLocationAdapter;
+    protected final JsonAdapter<ProbeLocation> probeLocationAdapter;
 
     public ProbeDetailsAdapter(Moshi moshi) {
-      probeLocationAdapter = moshi.adapter(Snapshot.ProbeLocation.class);
+      probeLocationAdapter = moshi.adapter(ProbeLocation.class);
     }
 
     @Override
