+++ dd-java-agent/agent-debugger/src/test/java/com/datadog/debugger/agent/CaptureAssertionHelper.java
@@ -4,7 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
-import datadog.trace.bootstrap.debugger.Snapshot;
+import com.datadog.debugger.sink.Snapshot;
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.List;
@@ -12,20 +13,20 @@ import java.util.Map;
 import utils.SourceCompiler;
 
 class CaptureAssertionHelper {
-  private static final Snapshot.CapturedThrowable HANDLED_EXCEPTION =
-      new Snapshot.CapturedThrowable(new IllegalArgumentException());
-  private static final Snapshot.CapturedThrowable UNHANDLED_EXCEPTION =
-      new Snapshot.CapturedThrowable(new RuntimeException());
+  private static final CapturedContext.CapturedThrowable HANDLED_EXCEPTION =
+      new CapturedContext.CapturedThrowable(new IllegalArgumentException());
+  private static final CapturedContext.CapturedThrowable UNHANDLED_EXCEPTION =
+      new CapturedContext.CapturedThrowable(new RuntimeException());
 
-  private final Map<DebuggerTransformerTest.ExceptionKind, Snapshot.CapturedValue[]> enterArgs =
-      new HashMap<>();
-  private final Map<DebuggerTransformerTest.ExceptionKind, Snapshot.CapturedValue[]> returnArgs =
-      new HashMap<>();
-  private final Map<DebuggerTransformerTest.ExceptionKind, Snapshot.CapturedValue> returnValues =
-      new HashMap<>();
-  private Snapshot.CapturedValue[] beforeLocals = null;
-  private Snapshot.CapturedValue[] afterLocals = null;
-  private Snapshot.CapturedValue[] returnLocals = null;
+  private final Map<DebuggerTransformerTest.ExceptionKind, CapturedContext.CapturedValue[]>
+      enterArgs = new HashMap<>();
+  private final Map<DebuggerTransformerTest.ExceptionKind, CapturedContext.CapturedValue[]>
+      returnArgs = new HashMap<>();
+  private final Map<DebuggerTransformerTest.ExceptionKind, CapturedContext.CapturedValue>
+      returnValues = new HashMap<>();
+  private CapturedContext.CapturedValue[] beforeLocals = null;
+  private CapturedContext.CapturedValue[] afterLocals = null;
+  private CapturedContext.CapturedValue[] returnLocals = null;
 
   private final DebuggerTransformerTest.InstrumentationKind instrumentationKind;
   private final SourceCompiler.DebugInfo debugInfo;
@@ -37,7 +38,7 @@ class CaptureAssertionHelper {
   private final String[] expectedArgNames;
   private final int expectedLineFrom;
   private final int expectedLineTill;
-  private final Snapshot.CapturedValue[] correlationFields;
+  private final CapturedContext.CapturedValue[] correlationFields;
   private final List<Snapshot> snapshots;
 
   CaptureAssertionHelper(
@@ -50,7 +51,7 @@ class CaptureAssertionHelper {
       Object returnValue,
       int expectedLineFrom,
       int expectedLineTill,
-      Snapshot.CapturedValue[] correlationFields,
+      CapturedContext.CapturedValue[] correlationFields,
       List<Snapshot> snapshots) {
     this.instrumentationKind = instrumentationKind;
     this.debugInfo = debugInfo;
@@ -82,9 +83,8 @@ class CaptureAssertionHelper {
 
   private void assertEntryExit(
       DebuggerTransformerTest.ExceptionKind exceptionKind, Snapshot.Captures captures) {
-    Snapshot.CapturedContext expectedEntryContext =
-        new Snapshot.CapturedContext(
-            enterArgs.get(exceptionKind), null, null, null, correlationFields);
+    CapturedContext expectedEntryContext =
+        new CapturedContext(enterArgs.get(exceptionKind), null, null, null, correlationFields);
     if (expectedLineFrom != -1) {
       // no entry/exit for line probe
       return;
@@ -93,8 +93,8 @@ class CaptureAssertionHelper {
     assertEquals(expectedEntryContext, captures.getEntry());
     if (exceptionKind != DebuggerTransformerTest.ExceptionKind.UNHANDLED) {
       // no return context is captured for unhandled exceptions
-      Snapshot.CapturedContext expectedExitContext =
-          new Snapshot.CapturedContext(
+      CapturedContext expectedExitContext =
+          new CapturedContext(
               returnArgs.get(exceptionKind),
               returnLocals,
               returnValues.get(exceptionKind),
@@ -107,7 +107,7 @@ class CaptureAssertionHelper {
     }
   }
 
-  private void addThis(Snapshot.CapturedContext expected, Snapshot.CapturedContext context) {
+  private void addThis(CapturedContext expected, CapturedContext context) {
     if (context.getArguments().containsKey("this")) {
       expected.getArguments().put("this", context.getArguments().get("this"));
     }
@@ -120,8 +120,8 @@ class CaptureAssertionHelper {
       return;
     }
     if (exceptionKind == DebuggerTransformerTest.ExceptionKind.UNHANDLED) {
-      Snapshot.CapturedContext expectedUnhandled =
-          new Snapshot.CapturedContext(
+      CapturedContext expectedUnhandled =
+          new CapturedContext(
               returnArgs.get(exceptionKind), null, null, UNHANDLED_EXCEPTION, correlationFields);
       assertEquals(
           expectedUnhandled.getThrowable().getMessage(),
@@ -146,16 +146,16 @@ class CaptureAssertionHelper {
     boolean isSingleline = expectedLineFrom == expectedLineTill;
     if (instrumentationKind != DebuggerTransformerTest.InstrumentationKind.ENTRY_EXIT
         && hasLineInfo(debugInfo)) {
-      Snapshot.CapturedContext expectedBefore =
-          new Snapshot.CapturedContext(
+      CapturedContext expectedBefore =
+          new CapturedContext(
               enterArgs.get(exceptionKind), beforeLocals, null, null, correlationFields);
       assertEquals(expectedBefore, captures.getLines().get(expectedLineFrom));
 
       if (!isSingleline
           && (exceptionKind == DebuggerTransformerTest.ExceptionKind.NONE
               || instrumentationKind == DebuggerTransformerTest.InstrumentationKind.LINE)) {
-        Snapshot.CapturedContext expectedAfter =
-            new Snapshot.CapturedContext(
+        CapturedContext expectedAfter =
+            new CapturedContext(
                 enterArgs.get(exceptionKind), afterLocals, null, null, correlationFields);
         assertEquals(expectedAfter, captures.getLines().get(expectedLineTill));
       } else {
@@ -176,7 +176,8 @@ class CaptureAssertionHelper {
         case HANDLED:
         case NONE:
           {
-            returnValues.put(exceptionKind, Snapshot.CapturedValue.of(returnType, returnValue));
+            returnValues.put(
+                exceptionKind, CapturedContext.CapturedValue.of(returnType, returnValue));
             break;
           }
       }
@@ -189,26 +190,26 @@ class CaptureAssertionHelper {
         case LINE:
           {
             returnLocals =
-                new Snapshot.CapturedValue[] {
-                  Snapshot.CapturedValue.of(
+                new CapturedContext.CapturedValue[] {
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.VAR_NAME, argumentType, argumentOutputValue)
                 };
             beforeLocals =
-                new Snapshot.CapturedValue[] {
-                  Snapshot.CapturedValue.of(
+                new CapturedContext.CapturedValue[] {
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.SCOPED_VAR_NAME,
                       DebuggerTransformerTest.SCOPED_VAR_TYPE,
                       DebuggerTransformerTest.SCOPED_VAR_VALUE),
-                  Snapshot.CapturedValue.of(
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.VAR_NAME, argumentType, argumentInputValue)
                 };
             afterLocals =
-                new Snapshot.CapturedValue[] {
-                  Snapshot.CapturedValue.of(
+                new CapturedContext.CapturedValue[] {
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.SCOPED_VAR_NAME,
                       DebuggerTransformerTest.SCOPED_VAR_TYPE,
                       DebuggerTransformerTest.SCOPED_VAR_VALUE),
-                  Snapshot.CapturedValue.of(
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.VAR_NAME, argumentType, argumentOutputValue)
                 };
             break;
@@ -216,18 +217,18 @@ class CaptureAssertionHelper {
         case ENTRY_EXIT:
           {
             returnLocals =
-                new Snapshot.CapturedValue[] {
-                  Snapshot.CapturedValue.of(
+                new CapturedContext.CapturedValue[] {
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.VAR_NAME, argumentType, argumentOutputValue)
                 };
             beforeLocals =
-                new Snapshot.CapturedValue[] {
-                  Snapshot.CapturedValue.of(
+                new CapturedContext.CapturedValue[] {
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.VAR_NAME, argumentType, argumentInputValue)
                 };
             afterLocals =
-                new Snapshot.CapturedValue[] {
-                  Snapshot.CapturedValue.of(
+                new CapturedContext.CapturedValue[] {
+                  CapturedContext.CapturedValue.of(
                       DebuggerTransformerTest.VAR_NAME, argumentType, argumentOutputValue)
                 };
             break;
@@ -241,23 +242,23 @@ class CaptureAssertionHelper {
         EnumSet.allOf(DebuggerTransformerTest.ExceptionKind.class)) {
       enterArgs.put(
           exceptionKind,
-          new Snapshot.CapturedValue[] {
-            Snapshot.CapturedValue.of(expectedArgNames[0], argumentType, argumentInputValue),
-            Snapshot.CapturedValue.of(expectedArgNames[1], "int", exceptionKind.ordinal())
+          new CapturedContext.CapturedValue[] {
+            CapturedContext.CapturedValue.of(expectedArgNames[0], argumentType, argumentInputValue),
+            CapturedContext.CapturedValue.of(expectedArgNames[1], "int", exceptionKind.ordinal())
           });
       returnArgs.put(
           exceptionKind,
-          new Snapshot.CapturedValue[] {
+          new CapturedContext.CapturedValue[] {
             // the first argument will not get mutated if an exception is thrown before reaching the
             // line of code
-            Snapshot.CapturedValue.of(
+            CapturedContext.CapturedValue.of(
                 expectedArgNames[0],
                 argumentType,
                 exceptionKind == DebuggerTransformerTest.ExceptionKind.NONE
                     ? argumentOutputValue
                     : argumentInputValue),
             // exception handler will overwrite the second argument with value of -1
-            Snapshot.CapturedValue.of(
+            CapturedContext.CapturedValue.of(
                 expectedArgNames[1],
                 "int",
                 exceptionKind == DebuggerTransformerTest.ExceptionKind.HANDLED
