+++ dd-java-agent/agent-debugger/src/test/java/com/datadog/debugger/agent/MetricProbesInstrumentationTest.java
@@ -9,13 +9,14 @@ import static utils.InstrumentationTestHelper.compileAndLoadClass;
 
 import com.datadog.debugger.el.DSL;
 import com.datadog.debugger.el.ValueScript;
+import com.datadog.debugger.instrumentation.DiagnosticMessage;
 import com.datadog.debugger.probe.MetricProbe;
+import com.datadog.debugger.sink.Sink;
+import com.datadog.debugger.sink.Snapshot;
 import datadog.trace.api.Config;
 import datadog.trace.bootstrap.debugger.DebuggerContext;
-import datadog.trace.bootstrap.debugger.DiagnosticMessage;
 import datadog.trace.bootstrap.debugger.MethodLocation;
 import datadog.trace.bootstrap.debugger.ProbeId;
-import datadog.trace.bootstrap.debugger.Snapshot;
 import java.io.IOException;
 import java.lang.instrument.ClassFileTransformer;
 import java.lang.instrument.Instrumentation;
@@ -747,7 +748,8 @@ public class MetricProbesInstrumentationTest {
     instr.addTransformer(currentTransformer);
     MetricForwarderListener listener = new MetricForwarderListener();
     mockSink = new MockSink();
-    DebuggerContext.init(mockSink, null, listener);
+    DebuggerAgentHelper.injectSink(mockSink);
+    DebuggerContext.init(null, listener);
     DebuggerContext.initClassFilter(new DenyListHelper(null));
     return listener;
   }
@@ -781,12 +783,15 @@ public class MetricProbesInstrumentationTest {
     }
   }
 
-  private static class MockSink implements DebuggerContext.Sink {
+  private static class MockSink implements Sink {
 
     private final List<DiagnosticMessage> currentDiagnostics = new ArrayList<>();
 
     @Override
     public void addSnapshot(Snapshot snapshot) {}
+
+    @Override
+    public void skipSnapshot(String probeId, DebuggerContext.SkipCause cause) {}
 
     @Override
     public void addDiagnostics(ProbeId probeId, List<DiagnosticMessage> messages) {
