+++ dd-java-agent/agent-debugger/src/test/java/com/datadog/debugger/sink/DebuggerSinkTest.java
@@ -10,8 +10,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
+import com.datadog.debugger.agent.DebuggerAgentHelper;
 import com.datadog.debugger.agent.JsonSnapshotSerializer;
 import com.datadog.debugger.agent.ProbeStatus;
+import com.datadog.debugger.instrumentation.DiagnosticMessage;
 import com.datadog.debugger.uploader.BatchUploader;
 import com.datadog.debugger.util.DebuggerMetrics;
 import com.datadog.debugger.util.MoshiHelper;
@@ -19,12 +21,13 @@ import com.datadog.debugger.util.MoshiSnapshotTestHelper;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Types;
 import datadog.trace.api.Config;
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.CapturedStackFrame;
 import datadog.trace.bootstrap.debugger.DebuggerContext;
-import datadog.trace.bootstrap.debugger.DiagnosticMessage;
+import datadog.trace.bootstrap.debugger.EvaluationError;
 import datadog.trace.bootstrap.debugger.ProbeId;
 import datadog.trace.bootstrap.debugger.ProbeImplementation;
-import datadog.trace.bootstrap.debugger.Snapshot;
+import datadog.trace.bootstrap.debugger.ProbeLocation;
 import java.io.IOException;
 import java.lang.reflect.ParameterizedType;
 import java.net.URISyntaxException;
@@ -43,8 +46,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 @ExtendWith(MockitoExtension.class)
 public class DebuggerSinkTest {
   private static final String PROBE_ID = "12fd-8490-c111-4374-ffde";
-  private static final Snapshot.ProbeLocation PROBE_LOCATION =
-      new Snapshot.ProbeLocation("java.lang.String", "indexOf", null, null);
+  private static final ProbeLocation PROBE_LOCATION =
+      new ProbeLocation("java.lang.String", "indexOf", null, null);
   public static final int MAX_PAYLOAD = 5 * 1024 * 1024;
 
   @Mock private Config config;
@@ -55,7 +58,7 @@ public class DebuggerSinkTest {
 
   @BeforeEach
   void setUp() {
-    DebuggerContext.initSnapshotSerializer(new JsonSnapshotSerializer());
+    DebuggerContext.initValueSerializer(new JsonSnapshotSerializer());
     when(config.getHostName()).thenReturn("host-name");
     when(config.getServiceName()).thenReturn("service-name");
     when(config.getEnv()).thenReturn("test");
@@ -71,6 +74,7 @@ public class DebuggerSinkTest {
   @Test
   public void addSnapshot() throws URISyntaxException, IOException {
     DebuggerSink sink = new DebuggerSink(config, batchUploader);
+    DebuggerAgentHelper.injectSerializer(new JsonSnapshotSerializer());
     Snapshot snapshot =
         new Snapshot(
             Thread.currentThread(),
@@ -94,6 +98,7 @@ public class DebuggerSinkTest {
   public void addMultipleSnapshots() throws URISyntaxException, IOException {
     when(config.getDebuggerUploadBatchSize()).thenReturn(2);
     DebuggerSink sink = new DebuggerSink(config, batchUploader);
+    DebuggerAgentHelper.injectSerializer(new JsonSnapshotSerializer());
     Snapshot snapshot =
         new Snapshot(
             Thread.currentThread(),
@@ -115,6 +120,7 @@ public class DebuggerSinkTest {
   public void splitSnapshotBatch() {
     when(config.getDebuggerUploadBatchSize()).thenReturn(10);
     DebuggerSink sink = new DebuggerSink(config, batchUploader);
+    DebuggerAgentHelper.injectSerializer(new JsonSnapshotSerializer());
     Snapshot largeSnapshot =
         new Snapshot(
             Thread.currentThread(),
@@ -310,11 +316,12 @@ public class DebuggerSinkTest {
   @Test
   public void addSnapshotWithCorrelationIdsMethodProbe() throws URISyntaxException, IOException {
     DebuggerSink sink = new DebuggerSink(config, batchUploader);
-    Snapshot.CapturedContext entry = new Snapshot.CapturedContext();
+    DebuggerAgentHelper.injectSerializer(new JsonSnapshotSerializer());
+    CapturedContext entry = new CapturedContext();
     entry.addFields(
-        new Snapshot.CapturedValue[] {
-          Snapshot.CapturedValue.of("dd.trace_id", "java.lang.String", "123"),
-          Snapshot.CapturedValue.of("dd.span_id", "java.lang.String", "456"),
+        new CapturedContext.CapturedValue[] {
+          CapturedContext.CapturedValue.of("dd.trace_id", "java.lang.String", "123"),
+          CapturedContext.CapturedValue.of("dd.span_id", "java.lang.String", "456"),
         });
     Snapshot snapshot =
         new Snapshot(
@@ -334,11 +341,12 @@ public class DebuggerSinkTest {
   @Test
   public void addSnapshotWithCorrelationIdsLineProbe() throws URISyntaxException, IOException {
     DebuggerSink sink = new DebuggerSink(config, batchUploader);
-    Snapshot.CapturedContext line = new Snapshot.CapturedContext();
+    DebuggerAgentHelper.injectSerializer(new JsonSnapshotSerializer());
+    CapturedContext line = new CapturedContext();
     line.addFields(
-        new Snapshot.CapturedValue[] {
-          Snapshot.CapturedValue.of("dd.trace_id", "java.lang.String", "123"),
-          Snapshot.CapturedValue.of("dd.span_id", "java.lang.String", "456"),
+        new CapturedContext.CapturedValue[] {
+          CapturedContext.CapturedValue.of("dd.trace_id", "java.lang.String", "123"),
+          CapturedContext.CapturedValue.of("dd.span_id", "java.lang.String", "456"),
         });
     Snapshot snapshot =
         new Snapshot(
@@ -358,21 +366,21 @@ public class DebuggerSinkTest {
   @Test
   public void addSnapshotWithEvalErrors() throws URISyntaxException, IOException {
     DebuggerSink sink = new DebuggerSink(config, batchUploader);
-    DebuggerContext.init(sink, null, null);
-    Snapshot.CapturedContext entry = new Snapshot.CapturedContext();
+    DebuggerAgentHelper.injectSerializer(new JsonSnapshotSerializer());
+    CapturedContext entry = new CapturedContext();
     ProbeImplementation probeImplementation =
         new ProbeImplementation.NoopProbeImplementation(PROBE_ID, PROBE_LOCATION);
     Snapshot snapshot = new Snapshot(Thread.currentThread(), probeImplementation);
     snapshot.setEntry(entry);
     snapshot.addEvaluationErrors(
-        Arrays.asList(new Snapshot.EvaluationError("obj.field", "Cannot dereference obj")));
+        Arrays.asList(new EvaluationError("obj.field", "Cannot dereference obj")));
     sink.addSnapshot(snapshot);
     sink.flush(sink);
     verify(batchUploader).upload(payloadCaptor.capture(), matches(EXPECTED_SNAPSHOT_TAGS));
     String strPayload = new String(payloadCaptor.getValue(), StandardCharsets.UTF_8);
     System.out.println(strPayload);
     JsonSnapshotSerializer.IntakeRequest intakeRequest = assertOneIntakeRequest(strPayload);
-    List<Snapshot.EvaluationError> evaluationErrors =
+    List<EvaluationError> evaluationErrors =
         intakeRequest.getDebugger().getSnapshot().getEvaluationErrors();
     assertEquals(1, evaluationErrors.size());
     assertEquals("obj.field", evaluationErrors.get(0).getExpr());
