+++ dd-java-agent/agent-debugger/src/test/java/com/datadog/debugger/util/MoshiSnapshotTestHelper.java
@@ -20,12 +20,14 @@ import static com.datadog.debugger.util.MoshiSnapshotHelper.TYPE;
 import static com.datadog.debugger.util.MoshiSnapshotHelper.VALUE;
 
 import com.datadog.debugger.el.ProbeCondition;
+import com.datadog.debugger.sink.Snapshot;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.ProbeImplementation;
-import datadog.trace.bootstrap.debugger.Snapshot;
+import datadog.trace.bootstrap.debugger.ProbeLocation;
 import datadog.trace.bootstrap.debugger.el.DebuggerScript;
 import java.io.IOException;
 import java.lang.annotation.Annotation;
@@ -46,10 +48,10 @@ public class MoshiSnapshotTestHelper {
         return new MoshiSnapshotTestHelper.CapturesAdapter(
             moshi, new CapturedContextAdapter(moshi, new CapturedValueAdapter()));
       }
-      if (Types.equals(type, Snapshot.CapturedValue.class)) {
+      if (Types.equals(type, CapturedContext.CapturedValue.class)) {
         return new MoshiSnapshotTestHelper.CapturedValueAdapter();
       }
-      if (Types.equals(type, Snapshot.CapturedContext.class)) {
+      if (Types.equals(type, CapturedContext.class)) {
         return new MoshiSnapshotTestHelper.CapturedContextAdapter(
             moshi, new CapturedValueAdapter());
       }
@@ -71,8 +73,7 @@ public class MoshiSnapshotTestHelper {
 
   private static class CapturesAdapter extends MoshiSnapshotHelper.CapturesAdapter {
 
-    public CapturesAdapter(
-        Moshi moshi, JsonAdapter<Snapshot.CapturedContext> capturedContextAdapter) {
+    public CapturesAdapter(Moshi moshi, JsonAdapter<CapturedContext> capturedContextAdapter) {
       super(moshi, capturedContextAdapter);
     }
 
@@ -90,13 +91,13 @@ public class MoshiSnapshotTestHelper {
             captures.setReturn(capturedContextAdapter.fromJson(jsonReader));
             break;
           case LINES:
-            Map<Integer, Snapshot.CapturedContext> map = linesAdapter.fromJson(jsonReader);
+            Map<Integer, CapturedContext> map = linesAdapter.fromJson(jsonReader);
             if (map != null) {
               map.forEach(captures::addLine);
             }
             break;
           case CAUGHT_EXCEPTIONS:
-            List<Snapshot.CapturedThrowable> capturedThrowables =
+            List<CapturedContext.CapturedThrowable> capturedThrowables =
                 caughtExceptionsAdapter.fromJson(jsonReader);
             capturedThrowables.forEach(captures::addCaughtException);
             break;
@@ -110,23 +111,24 @@ public class MoshiSnapshotTestHelper {
   }
 
   public static class CapturedContextAdapter extends MoshiSnapshotHelper.CapturedContextAdapter {
-    private static final Snapshot.CapturedValue[] EMPTY_VALUES_ARRAY =
-        new Snapshot.CapturedValue[0];
+    private static final CapturedContext.CapturedValue[] EMPTY_VALUES_ARRAY =
+        new CapturedContext.CapturedValue[0];
 
-    public CapturedContextAdapter(Moshi moshi, JsonAdapter<Snapshot.CapturedValue> valueAdapter) {
+    public CapturedContextAdapter(
+        Moshi moshi, JsonAdapter<CapturedContext.CapturedValue> valueAdapter) {
       super(moshi, valueAdapter);
     }
 
     @Override
-    public Snapshot.CapturedContext fromJson(JsonReader jsonReader) throws IOException {
+    public CapturedContext fromJson(JsonReader jsonReader) throws IOException {
       jsonReader.beginObject();
-      Snapshot.CapturedContext capturedContext = new Snapshot.CapturedContext();
+      CapturedContext capturedContext = new CapturedContext();
       while (jsonReader.hasNext()) {
         String name = jsonReader.nextName();
         switch (name) {
           case ARGUMENTS:
             jsonReader.beginObject();
-            List<Snapshot.CapturedValue> argValues = new ArrayList<>();
+            List<CapturedContext.CapturedValue> argValues = new ArrayList<>();
             while (jsonReader.hasNext()) {
               String argName = jsonReader.peekJson().nextName();
               if ("this".equals(argName)) {
@@ -135,7 +137,7 @@ public class MoshiSnapshotTestHelper {
                 continue;
               }
               argName = jsonReader.nextName();
-              Snapshot.CapturedValue capturedValue = valueAdapter.fromJson(jsonReader);
+              CapturedContext.CapturedValue capturedValue = valueAdapter.fromJson(jsonReader);
               if (capturedValue != null) {
                 capturedValue.setName(argName);
                 argValues.add(capturedValue);
@@ -158,7 +160,7 @@ public class MoshiSnapshotTestHelper {
       return capturedContext;
     }
 
-    private void fromJsonFields(JsonReader jsonReader, Snapshot.CapturedContext capturedContext)
+    private void fromJsonFields(JsonReader jsonReader, CapturedContext capturedContext)
         throws IOException {
       jsonReader.beginObject();
       while (jsonReader.hasNext()) {
@@ -181,17 +183,17 @@ public class MoshiSnapshotTestHelper {
       jsonReader.endObject();
     }
 
-    private Snapshot.CapturedValue[] fromJsonCapturedValues(JsonReader jsonReader)
+    private CapturedContext.CapturedValue[] fromJsonCapturedValues(JsonReader jsonReader)
         throws IOException {
       jsonReader.beginObject();
-      List<Snapshot.CapturedValue> values = new ArrayList<>();
+      List<CapturedContext.CapturedValue> values = new ArrayList<>();
       while (jsonReader.hasNext()) {
         String name = jsonReader.nextName();
         if (NOT_CAPTURED_REASON.equals(name)) {
           jsonReader.nextString();
           continue;
         }
-        Snapshot.CapturedValue capturedValue = valueAdapter.fromJson(jsonReader);
+        CapturedContext.CapturedValue capturedValue = valueAdapter.fromJson(jsonReader);
         if (capturedValue != null) {
           capturedValue.setName(name);
           values.add(capturedValue);
@@ -204,7 +206,7 @@ public class MoshiSnapshotTestHelper {
 
   public static class CapturedValueAdapter extends MoshiSnapshotHelper.CapturedValueAdapter {
     @Override
-    public Snapshot.CapturedValue fromJson(JsonReader jsonReader) throws IOException {
+    public CapturedContext.CapturedValue fromJson(JsonReader jsonReader) throws IOException {
       jsonReader.beginObject();
       String type = null;
       Object value = null;
@@ -221,14 +223,14 @@ public class MoshiSnapshotTestHelper {
             break;
           case FIELDS:
             jsonReader.beginObject();
-            Map<String, Snapshot.CapturedValue> fields = new HashMap<>();
+            Map<String, CapturedContext.CapturedValue> fields = new HashMap<>();
             while (jsonReader.hasNext()) {
               String fieldName = jsonReader.nextName();
               if (NOT_CAPTURED_REASON.equals(fieldName)) {
                 notCapturedReason = jsonReader.nextString();
                 continue;
               }
-              Snapshot.CapturedValue fieldValue = fromJson(jsonReader);
+              CapturedContext.CapturedValue fieldValue = fromJson(jsonReader);
               fields.put(fieldName, fieldValue);
             }
             jsonReader.endObject();
@@ -240,9 +242,9 @@ public class MoshiSnapshotTestHelper {
                 throw new RuntimeException("type is null");
               }
               jsonReader.beginArray();
-              List<Snapshot.CapturedValue> values = new ArrayList<>();
+              List<CapturedContext.CapturedValue> values = new ArrayList<>();
               while (jsonReader.hasNext()) {
-                Snapshot.CapturedValue elementValue = fromJson(jsonReader);
+                CapturedContext.CapturedValue elementValue = fromJson(jsonReader);
                 values.add(elementValue);
               }
               jsonReader.endArray();
@@ -250,7 +252,7 @@ public class MoshiSnapshotTestHelper {
                   || type.equals(ArrayList.class.getTypeName())
                   || type.equals("java.util.Collections$UnmodifiableRandomAccessList")) {
                 List<Object> list = new ArrayList<>();
-                for (Snapshot.CapturedValue cValue : values) {
+                for (CapturedContext.CapturedValue cValue : values) {
                   list.add(cValue.getValue());
                 }
                 value = list;
@@ -259,7 +261,7 @@ public class MoshiSnapshotTestHelper {
                 if (SerializerWithLimits.isPrimitive(componentType)) {
                   value = createPrimitiveArray(componentType, values);
                 } else {
-                  value = values.stream().map(Snapshot.CapturedValue::getValue).toArray();
+                  value = values.stream().map(CapturedContext.CapturedValue::getValue).toArray();
                 }
               } else if (type.equals("java.util.Collections$EmptyList")) {
                 value = Collections.emptyList();
@@ -271,10 +273,10 @@ public class MoshiSnapshotTestHelper {
           case ENTRIES:
             {
               jsonReader.beginArray();
-              List<Snapshot.CapturedValue> values = new ArrayList<>();
+              List<CapturedContext.CapturedValue> values = new ArrayList<>();
               while (jsonReader.hasNext()) {
                 jsonReader.beginArray();
-                Snapshot.CapturedValue elementValue = fromJson(jsonReader);
+                CapturedContext.CapturedValue elementValue = fromJson(jsonReader);
                 values.add(elementValue);
                 elementValue = fromJson(jsonReader);
                 values.add(elementValue);
@@ -314,16 +316,17 @@ public class MoshiSnapshotTestHelper {
         }
       }
       jsonReader.endObject();
-      return Snapshot.CapturedValue.raw(type, value, notCapturedReason);
+      return CapturedContext.CapturedValue.raw(type, value, notCapturedReason);
     }
 
-    private Object createPrimitiveArray(String componentType, List<Snapshot.CapturedValue> values) {
+    private Object createPrimitiveArray(
+        String componentType, List<CapturedContext.CapturedValue> values) {
       switch (componentType) {
         case "byte":
           {
             byte[] bytes = new byte[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               bytes[i++] = (Byte) capturedValue.getValue();
             }
             return bytes;
@@ -332,7 +335,7 @@ public class MoshiSnapshotTestHelper {
           {
             boolean[] booleans = new boolean[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               booleans[i++] = (Boolean) capturedValue.getValue();
             }
             return booleans;
@@ -341,7 +344,7 @@ public class MoshiSnapshotTestHelper {
           {
             short[] shorts = new short[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               shorts[i++] = (Short) capturedValue.getValue();
             }
             return shorts;
@@ -350,7 +353,7 @@ public class MoshiSnapshotTestHelper {
           {
             char[] chars = new char[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               chars[i++] = (Character) capturedValue.getValue();
             }
             return chars;
@@ -359,7 +362,7 @@ public class MoshiSnapshotTestHelper {
           {
             int[] ints = new int[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               ints[i++] = (Integer) capturedValue.getValue();
             }
             return ints;
@@ -368,7 +371,7 @@ public class MoshiSnapshotTestHelper {
           {
             long[] longs = new long[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               longs[i++] = (Long) capturedValue.getValue();
             }
             return longs;
@@ -377,7 +380,7 @@ public class MoshiSnapshotTestHelper {
           {
             float[] floats = new float[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               floats[i++] = (Float) capturedValue.getValue();
             }
             return floats;
@@ -386,7 +389,7 @@ public class MoshiSnapshotTestHelper {
           {
             double[] doubles = new double[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               doubles[i++] = (Double) capturedValue.getValue();
             }
             return doubles;
@@ -395,7 +398,7 @@ public class MoshiSnapshotTestHelper {
           {
             String[] strings = new String[values.size()];
             int i = 0;
-            for (Snapshot.CapturedValue capturedValue : values) {
+            for (CapturedContext.CapturedValue capturedValue : values) {
               strings[i++] = (String) capturedValue.getValue();
             }
             return strings;
@@ -450,7 +453,7 @@ public class MoshiSnapshotTestHelper {
     @Override
     public ProbeImplementation fromJson(JsonReader jsonReader) throws IOException {
       String id = null;
-      Snapshot.ProbeLocation location = null;
+      ProbeLocation location = null;
       jsonReader.beginObject();
       while (jsonReader.hasNext()) {
         String name = jsonReader.nextName();
