+++ dd-java-agent/agent-debugger/src/test/java/com/datadog/debugger/util/ValueScriptHelperTest.java
@@ -2,8 +2,9 @@ package com.datadog.debugger.util;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import datadog.trace.bootstrap.debugger.CapturedContext;
 import datadog.trace.bootstrap.debugger.ProbeImplementation;
-import datadog.trace.bootstrap.debugger.Snapshot;
+import datadog.trace.bootstrap.debugger.ProbeLocation;
 import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
@@ -11,15 +12,14 @@ class ValueScriptHelperTest {
 
   private static final String PROBE_ID = "12fd-8490-c111-4374-ffde";
   private static final int PROBE_VERSION = 42;
-  private static final Snapshot.ProbeLocation PROBE_LOCATION =
-      new Snapshot.ProbeLocation(
-          "java.lang.String", "indexOf", "String.java", Arrays.asList("12-15", "23"));
+  private static final ProbeLocation PROBE_LOCATION =
+      new ProbeLocation("java.lang.String", "indexOf", "String.java", Arrays.asList("12-15", "23"));
   private static final ProbeImplementation DUMMY_PROBE =
       new ProbeImplementation.NoopProbeImplementation(PROBE_ID, PROBE_LOCATION);
 
   @Test
   public void nullValue() {
-    Snapshot.CapturedContext.Status status = new Snapshot.CapturedContext.Status(DUMMY_PROBE);
+    CapturedContext.Status status = new CapturedContext.Status(DUMMY_PROBE);
     StringBuilder sb = new StringBuilder();
     ValueScriptHelper.serializeValue(sb, "", null, status);
     assertEquals("null", sb.toString());
@@ -27,7 +27,7 @@ class ValueScriptHelperTest {
 
   @Test
   public void basicString() {
-    Snapshot.CapturedContext.Status status = new Snapshot.CapturedContext.Status(DUMMY_PROBE);
+    CapturedContext.Status status = new CapturedContext.Status(DUMMY_PROBE);
     StringBuilder sb = new StringBuilder();
     ValueScriptHelper.serializeValue(sb, "", "foo", status);
     assertEquals("foo", sb.toString());
